+++ attributes/docker.rb
@@ -1,0 +1,7 @@
+
+default['drone']['docker_tls_verify']  = true
+default['drone']['docker_tls_ca_crt']  = nil
+default['drone']['docker_tls_server_crt']  = nil
+default['drone']['docker_tls_server_key']  = nil
+default['drone']['docker_tls_client_crt']  = nil
+default['drone']['docker_tls_client_key']  = nil