+++ attributes/integrations.rb
@@ -1,0 +1,3 @@
+
+# generate self-signed certs?
+default['drone']['generate_certs'] = false